rename getViewsFromConfig to getDocumentView
